Reorganize classes for flexibility
