+++ DupuitLEM/__init__.py
@@ -1,10 +1,3 @@
 from .models.simple_steady_model import SimpleSteadyRecharge
 from .models.stochastic_shear_stress_model import StochasticRechargeShearStress
 from .models.steady_shear_stress_model import SteadyRechargeShearStress
-
-from .grid_functions.grid_funcs import (
-    bind_avg_hydraulic_conductivity,
-    calc_shear_stress_manning,
-    calc_shear_stress_chezy,
-    calc_erosion_from_shear_stress
-    )
