+++ DupuitLEM/grid_functions/__init__.py
@@ -1,6 +1,6 @@
 from .grid_funcs import (
     bind_avg_hydraulic_conductivity,
-    calc_shear_stress_manning,
-    calc_shear_stress_chezy,
-    calc_erosion_from_shear_stress
+    bind_shear_stress_manning,
+    bind_shear_stress_chezy,
+    bind_erosion_from_shear_stress
     )
